requirements.txt
a2a-sdk==0.3.5
langchain-core==0.3.75
langgraph==0.6.4
fastapi==0.116.1
python-dotenv==1.1.1